refactor($icon): use cube for cascader and data picker
