+++ components/cascader/style/index.less
@@ -74,12 +74,12 @@
       top: 50%;
       right: @control-padding-horizontal;
       z-index: 2;
-      width: 12px;
-      height: 12px;
-      margin-top: -6px;
+      width: 16px;
+      height: 16px;
+      margin-top: -8px;
       color: @disabled-color;
       font-size: @font-size-sm;
-      line-height: 12px;
+      line-height: 16px;
       background: @component-background;
       cursor: pointer;
       opacity: 0;
