+++ components/cascader/style/index.ts
@@ -6,3 +6,4 @@ import '../../empty/style';
 import '../../input/style';
 import '../../spin/style';
 import '../../basic-arrow/style';
+import '../../basic-clear/style';
