+++ components/date-picker/style/index.tsx
@@ -3,3 +3,4 @@ import './index.less';
 // style dependencies
 import '../../tag/style';
 import '../../button/style';
+import '../../basic-clear/style';
